requirements.txt
matplotlib
numpy
Pillow
PyWavelets
scikit-image
scipy
networkx
oct2py